Correção do Método PUT e implementação de Exceptions

+++ src/main/java/br/start/petshop/services/PetService.java
@@ -2,9 +2,13 @@ package br.start.petshop.services;
 
 import br.start.petshop.DTOs.PetDTO;
 import br.start.petshop.entities.Pet;
+import br.start.petshop.exceptions.NotFoundException;
+import br.start.petshop.exceptions.NullException;
 import br.start.petshop.repositories.PetRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -31,21 +35,39 @@ public class PetService {
 
     public Pet save(PetDTO dto) {
         Pet pet = toEntity(dto);
+        if(pet != null) {
         return repo.save(pet);
+        }
+        else throw new NullException("Pet is null");
     }
 
     public Pet update(Long id, PetDTO dto) {
-        Pet pet = repo.findById(id).orElse(null);
-        if (pet != null) {
-        pet = toEntity(dto);
-        return repo.save(pet);
-        }
+        Optional<Pet> optionalPet = repo.findById(id);
+        if (optionalPet.isPresent()) {
+            Pet pet = optionalPet.get();
+            pet.setName(dto.name());
+            pet.setBirthDate(dto.birthDate());
+            pet.setGender(dto.gender());
+            pet.setSpecies(dto.species());
+            pet.setBreed(dto.breed());
+            pet.setSize(dto.size());
+            pet.setServiceType(dto.serviceType());
 
-        return null;
+            return repo.save(pet);
+        } else {
+        throw new NotFoundException("Pet not found with id: " + id);
+        }
     }
 
     public void delete(final Long id) {
-        repo.findById(id);
+
+        Optional<Pet> pet = repo.findById(id);
+        if (pet.isPresent()) {
+
         repo.deleteById(id);
+        }
+        else {
+            throw new NotFoundException("Pet not found with id: " + id);
+        }
     }
 }
